fix(labels): Better colours for light mode
Adjusted HSL values for text and background to find a happy medium
between light and dark mode. Still not exactly the same as GH labels,
but at least more usable now. Need to consider using a color picker and
converting to HSL to see how GH calculates the colours of the text.

Should also see if they do something differently in light mode.